requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
qrcode==7.4.2
pillow==10.1.0